Move certificate-related type into new certificate module
