+++ examples/print-cert.rs
@@ -6,7 +6,7 @@ use std::io;
 use x509_parser::extensions::*;
 use x509_parser::objects::*;
 use x509_parser::pem::parse_x509_pem;
-use x509_parser::x509::X509Certificate;
+use x509_parser::X509Certificate;
 use x509_parser::{parse_x509_certificate, X509Extension};
 
 fn print_hex_dump(bytes: &[u8], max_len: usize) {
